memorystorage: refactor to make disttae not depend on memorystorage package
